+++ lib/src/usage_impl_io.dart
@@ -53,6 +53,9 @@ class IOPostHandler extends PostHandler {
       return req.close();
     }).then((HttpClientResponse response) {
       response.drain();
+    }).catchError((e) {
+      // Catch errors that can happen during a request, but that we can't do
+      // anything about, e.g. a missing internet conenction.
     });
   }
 }